Extract renderer duplicate utilities into focused modules
