+++ opennow-stable/src/renderer/src/components/ElapsedSessionIndicators.tsx
@@ -4,19 +4,9 @@ import { Calendar, Clock3, Clock } from "lucide-react";
 
 import type { SubscriptionInfo } from "@shared/gfn";
 
-import { formatRemainingPlaytimeFromSubscription } from "../utils/usePlaytime";
 import { useTranslation } from "../i18n";
-
-function formatElapsed(totalSeconds: number): string {
-  const safe = Math.max(0, Math.floor(totalSeconds));
-  const hours = Math.floor(safe / 3600);
-  const minutes = Math.floor((safe % 3600) / 60);
-  const seconds = safe % 60;
-  if (hours > 0) {
-    return `${hours}:${minutes.toString().padStart(2, "0")}:${seconds.toString().padStart(2, "0")}`;
-  }
-  return `${minutes}:${seconds.toString().padStart(2, "0")}`;
-}
+import { formatElapsed } from "../utils/timeFormat";
+import { formatRemainingPlaytimeFromSubscription } from "../utils/usePlaytime";
 
 interface SessionElapsedIndicatorProps {
   startedAtMs: number | null;
