+++ opennow-stable/src/renderer/src/components/controllerMode/controllerLibrary/genreHelpers.ts
@@ -1,0 +1,5 @@
+export function sanitizeGenreName(raw: string): string {
+  return raw
+    .replace(/_/g, " ")
+    .replace(/\b\w/g, (ch) => ch.toUpperCase());
+}
